Add new common function for registering exit handlers
